+++ include/ata.h
@@ -8,13 +8,13 @@
 #include <variant>
 #include <vector>
 
-#define ATA_VERSION "1.16.2"
+#define ATA_VERSION "1.17.0"
 
 namespace ata {
 
 inline constexpr uint32_t VERSION_MAJOR = 1;
-inline constexpr uint32_t VERSION_MINOR = 16;
-inline constexpr uint32_t VERSION_REVISION = 2;
+inline constexpr uint32_t VERSION_MINOR = 17;
+inline constexpr uint32_t VERSION_REVISION = 0;
 
 inline constexpr std::string_view version() noexcept {
   return ATA_VERSION;